+++ src/handlers/messagedelete.ts
@@ -1,4 +1,5 @@
-import { Colors, TextChannel } from 'discord.js';
+import { Colors, AuditLogEvent, TextChannel } from 'discord.js';
+import moment from 'moment';
 import { Channels, UserLogExclude } from '../constants';
 import { makeEmbed } from '../lib/embed';
 
@@ -17,24 +18,94 @@ module.exports = {
             return;
         }
 
+        const fetchedLogs = await message.guild.fetchAuditLogs({
+            limit: 1,
+            type: AuditLogEvent.MessageDelete,
+        });
+
+        const deletionLog = fetchedLogs.entries.first();
+
+        const currentDate = new Date();
+        const formattedDate: string = moment(currentDate).utcOffset(0).format('DD, MM, YYYY, HH:mm:ss');
+
         const userLogsChannel = message.guild.channels.resolve(Channels.USER_LOGS) as TextChannel | null;
 
-        if (userLogsChannel && !UserLogExclude.some((e) => e === message.author.id)) {
-            const messageDeleteEmbed = makeEmbed({
-                color: Colors.Red,
-                thumbnail: { url: 'https://cdn.discordapp.com/attachments/770835189419999262/779946282373873694/150-1509174_deleted-message-icon-sign-hd-png-download.png' },
-                author: {
-                    name: message.author.tag,
-                    iconURL: message.author.displayAvatarURL({ dynamic: true }),
+        const messageDeleteEmbedNoLog = (formattedDate) => makeEmbed({
+            color: Colors.Red,
+            thumbnail: { url: 'https://cdn.discordapp.com/attachments/770835189419999262/779946282373873694/150-1509174_deleted-message-icon-sign-hd-png-download.png' },
+            author: {
+                name: message.author.tag,
+                iconURL: message.author.displayAvatarURL({ dynamic: true }),
+            },
+            fields: [
+                {
+                    name: 'Author',
+                    value: `${message.author}`,
+                    inline: true,
                 },
-                fields: [
-                    { name: 'Author', value: `${message.author}`, inline: true },
-                    { name: 'Channel', value: `${message.channel}`, inline: true },
-                    { name: 'Deleted Message', value: message.content ? `${message.content}` : FEATURE_NOT_AVAIL, inline: false },
-                ],
-                footer: { text: `User ID: ${message.author.id}` },
-            });
-            await userLogsChannel.send({ embeds: [messageDeleteEmbed] });
+                {
+                    name: 'Channel',
+                    value: `${message.channel}`,
+                    inline: true,
+                },
+                {
+                    name: 'Deleted by',
+                    value: 'No audit log was found, message was either deleted by author, or a bot',
+                    inline: false,
+                },
+                {
+                    name: 'Deleted Message',
+                    value: message.content ? `${message.content}` : FEATURE_NOT_AVAIL,
+                    inline: false,
+                },
+                { name: 'Date', value: formattedDate, inline: false },
+            ],
+            footer: { text: `User ID: ${message.author.id}` },
+        });
+
+        const messageDeleteEmbed = (executor, formattedDate) => makeEmbed({
+            color: Colors.Red,
+            thumbnail: { url: 'https://cdn.discordapp.com/attachments/770835189419999262/779946282373873694/150-1509174_deleted-message-icon-sign-hd-png-download.png' },
+            author: {
+                name: message.author.tag,
+                iconURL: message.author.displayAvatarURL({ dynamic: true }),
+            },
+            fields: [
+                {
+                    name: 'Author',
+                    value: `${message.author}`,
+                    inline: true,
+                },
+                {
+                    name: 'Channel',
+                    value: `${message.channel}`,
+                    inline: true,
+                },
+                {
+                    name: 'Deleted by',
+                    value: `${executor}`,
+                    inline: false,
+                },
+                {
+                    name: 'Deleted Message',
+                    value: message.content ? `${message.content}` : FEATURE_NOT_AVAIL,
+                    inline: false,
+                },
+                { name: 'Date', value: formattedDate, inline: false },
+            ],
+            footer: { text: `User ID: ${message.author.id}` },
+        });
+
+        if (userLogsChannel && !UserLogExclude.some((e) => e === message.author.id)) {
+            if (!deletionLog) await userLogsChannel.send({ embeds: [messageDeleteEmbedNoLog(formattedDate)] });
+
+            const { executor, target } = deletionLog;
+
+            if (target.id === message.author.id) {
+                await userLogsChannel.send({ embeds: [messageDeleteEmbed(executor, formattedDate)] });
+            } else {
+                await userLogsChannel.send({ embeds: [messageDeleteEmbedNoLog(formattedDate)] });
+            }
         }
     },
 };